refactor: :recycle: refactor kontakt department page
